Add Policy To Modules

+++ Modules/RolePermissions/Http/Controllers/RolePermissionsController.php
@@ -2,6 +2,7 @@
 
 namespace Modules\RolePermissions\Http\Controllers;
 
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
@@ -10,6 +11,7 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Modules\RolePermissions\Http\Requests\RolePermissionRequest;
+use Modules\RolePermissions\Models\Role;
 use Modules\RolePermissions\Repositories\RolePermissionsRepo;
 use Modules\RolePermissions\Services\RolePermissionsService;
 use Modules\Share\Http\Controllers\Controller;
@@ -31,9 +33,11 @@ class RolePermissionsController extends Controller
      * Display a listing of the resource.
      *
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('manage', Role::class);
         $roles = $this->repo->index()->paginate(10);
         return view('RolePermissions::index', compact('roles'));
     }
@@ -42,9 +46,11 @@ class RolePermissionsController extends Controller
      * Show the form for creating a new resource.
      *
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function create()
     {
+        $this->authorize('manage', Role::class);
         $permissions = $this->repo->getAllPermissions();
         return view('RolePermissions::create', compact('permissions'));
     }
@@ -54,9 +60,11 @@ class RolePermissionsController extends Controller
      *
      * @param RolePermissionRequest $request
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(RolePermissionRequest $request)
     {
+        $this->authorize('manage', Role::class);
         $this->service->store($request);
         return $this->successMessage('Roles Created Successfully');
     }
@@ -77,9 +85,11 @@ class RolePermissionsController extends Controller
      *
      * @param  int  $id
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function edit($id)
     {
+        $this->authorize('manage', Role::class);
         $role = $this->repo->findById($id);
         $permissions = $this->repo->getAllPermissions();
         return view('RolePermissions::edit', compact('role', 'permissions'));
@@ -89,11 +99,13 @@ class RolePermissionsController extends Controller
      * Update the specified resource in storage.
      *
      * @param Request $request
-     * @param  int  $id
+     * @param int $id
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(RolePermissionRequest $request, $id)
     {
+        $this->authorize('manage', Role::class);
         $this->service->update($request, $id);
         return $this->successMessage('Roles Updated Successfully');
     }
@@ -103,9 +115,11 @@ class RolePermissionsController extends Controller
      *
      * @param  int  $id
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('manage', Role::class);
         $this->repo->delete($id);
         return AjaxResponses::SuccessResponse();
     }
