+++ Modules/User/Http/Controllers/UserController.php
@@ -2,16 +2,20 @@
 
 namespace Modules\User\Http\Controllers;
 
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Modules\Share\Http\Controllers\Controller;
 use Modules\Share\Repositories\ShareRepo;
 use Modules\Share\Responses\AjaxResponses;
+use Modules\User\Http\Requests\AddRoleRequest;
 use Modules\User\Http\Requests\UserRequest;
+use Modules\User\Models\User;
 use Modules\User\Repositories\UserRepo;
 use Modules\User\Services\UserService;
 
@@ -30,9 +34,11 @@ class UserController extends Controller
      * Display a listing of the resource.
      *
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('manage', User::class);
         $users = $this->repo->index(auth()->id())->paginate(10);
         return view('User::index', compact('users'));
     }
@@ -41,9 +47,11 @@ class UserController extends Controller
      * Show the form for creating a new resource.
      *
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function create()
     {
+        $this->authorize('manage', User::class);
         return view('User::create');
     }
 
@@ -52,9 +60,11 @@ class UserController extends Controller
      *
      * @param UserRequest $request
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(UserRequest $request)
     {
+        $this->authorize('manage', User::class);
         $this->service->store($request);
         return $this->successMessage('User Created Successfully');
     }
@@ -75,9 +85,11 @@ class UserController extends Controller
      *
      * @param  int  $id
      * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function edit($id)
     {
+        $this->authorize('manage', User::class);
         $user = $this->repo->findById(id: $id);
         return view('User::edit', compact('user'));
     }
@@ -88,9 +100,11 @@ class UserController extends Controller
      * @param UserRequest $request
      * @param int $id
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(UserRequest $request, $id)
     {
+        $this->authorize('manage', User::class);
         $this->service->update($request, $id);
         return $this->successMessage('User Updated Successfully');
     }
@@ -98,13 +112,36 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\JsonResponse
+     * @param int $id
+     * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('manage', User::class);
         $this->repo->delete(id: $id);
         return AjaxResponses::SuccessResponse();
+    }
+
+    public function addRole(AddRoleRequest $request, User $user)
+    {
+        $this->authorize('manage', User::class);
+        $user->assignRole($request->role);
+        return $this->successMessage('Assign Role Successfully');
+    }
+
+    public function removeRole($userId, $role)
+    {
+        $this->authorize('manage', User::class);
+        $user = $this->repo->findById($userId);
+        $user->removeRole($role);
+        return AjaxResponses::SuccessResponse();
+    }
+
+    private function successMessage($text)
+    {
+        ShareRepo::successMessage(text: $text);
+        return redirect()->route('users.index');
     }
 
 //    public function changeStatusEmailVerified($id)
@@ -112,10 +149,4 @@ class UserController extends Controller
 //        $this->repo->changeStatusEmailVerified($id);
 //        return AjaxResponses::SuccessResponse();
 //    }
-
-    private function successMessage($text)
-    {
-        ShareRepo::successMessage(text: $text);
-        return redirect()->route('users.index');
-    }
 }
